Cleaned up code and refactored

+++ Open CV/src/main/java/org/opencv/imgproc/Imgproc.java
@@ -1644,7 +1644,7 @@ public class Imgproc {
     public static Point phaseCorrelate(Mat src1, Mat src2, Mat window, double[] response) {
         double[] response_out = new double[1];
         Point retVal = new Point(phaseCorrelate_0(src1.nativeObj, src2.nativeObj, window.nativeObj, response_out));
-        if (response != null) response[0] = (double) response_out[0];
+        if (response != null) response[0] = response_out[0];
         return retVal;
     }
